+++ PdfOcrReader/README.md
