+++ docker-compose.yml
@@ -1,0 +1,11 @@
+
+
+version: '3'
+services:
+    ubuntu:
+        build: .
+        container_name: 'ubuntu20.04'
+        working_dir: '/root/'
+        tty: true
+        volumes:
+            - ./PdfOcrReader:/root/PdfOcrReader
